signed/unsigned comparison
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@5925 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/ResourceEditor.cpp
@@ -633,7 +633,7 @@ CResourceDirectory* CResourceEditor::ScanDirectory(PRESOURCE_DIRECTORY rdRoot, P
 
       if (m_bKeepData)
       {
-        if (dwOffset > m_iSize)
+        if (dwOffset > DWORD(m_iSize))
         {
           throw runtime_error("Invalid resource entry data pointer, possibly compressed resources");
         }
